requirements.txt
streamlit
requests
beautifulsoup4
readability-lxml
chromadb
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
tqdm